HUGE update : - more dynamic strat (not everything work unfortunately) - beginning of an easy static strat - some refactoring - implementation of some basic mission

+++ src/modelec_com/include/modelec_com/pcb_odo_interface.hpp
@@ -39,7 +39,7 @@ public:
     struct OdometryData {
         long x;
         long y;
-        long theta;
+        double theta;
     };
 
     struct PIDData
